further fixes to primality check

+++ src/test/java/com/westpacnz/primenumbercalculator/CalculatorServiceTest.java
@@ -17,24 +17,24 @@ class CalculatorServiceTest {
 	@Test
 	void getPrimeNumberSum_givenCorrectArgs_returnPrimeSum() {
 		// the numbers added within expectedVal are the prime numbers before 10
-		Integer expectedVal = 2 + 3 + 5 + 7;
+		Long expectedVal = 2L + 3L + 5L + 7L;
 		assertEquals(expectedVal, calculatorService.getPrimeNumberSum(1L,10L));
 		assertEquals(expectedVal, calculatorService.getPrimeNumberSum(2L,10L));
 		assertEquals(expectedVal, calculatorService.getPrimeNumberSum(2L,7L));
 
 		// 11 is a prime number
-		expectedVal += 11;
+		expectedVal += 11L;
 		assertEquals(expectedVal, calculatorService.getPrimeNumberSum(1L,11L));
 	}
 
 	@Test
 	void getPrimeNumberSum_givenIncorrectArgs_return0() {
 		// startNum is greater than endNum
-		assertEquals(Integer.valueOf(0), calculatorService.getPrimeNumberSum(10L,9L));
+		assertEquals(Long.valueOf(0), calculatorService.getPrimeNumberSum(10L,9L));
 		// startNum is null
-		assertEquals(Integer.valueOf(0), calculatorService.getPrimeNumberSum(null,9L));
+		assertEquals(Long.valueOf(0), calculatorService.getPrimeNumberSum(null,9L));
 		// endNum is null
-		assertEquals(Integer.valueOf(0), calculatorService.getPrimeNumberSum(6L,null));
+		assertEquals(Long.valueOf(0), calculatorService.getPrimeNumberSum(6L,null));
 	}
 
 }
